rename module to `generalized_qsp`
